Update and rename 785. Is Graph Bipartite? to 785. Is Graph Bipartite?.java
Easy to judge no triangle situation, but need be careful about nodes that have no neighbors, in other way to say is that their graph[id] length is zero. So we need to loop through the whole graph arrays. It is similar to spanning tree or connected component in graph.